respaldo 17/03/07

+++ src/main/java/com/newtech/model/Pozo.java
@@ -32,7 +32,7 @@ public class Pozo implements Serializable {
 	@Column(name="id_pozo")
 	@GeneratedValue(generator="increment")
     @GenericGenerator(name="increment", strategy = "increment")
-	private Long idPozo;
+	private Integer idPozo;
 
 	@Column(name="desc_pozo")
 	private String descPozo;
